% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_propluvia.R
\encoding{UTF-8}
\name{get_propluvia}
\alias{get_propluvia}
\title{Télécharge l'ensemble des données Propluvia à date du jour ou spécifiée}
\usage{
get_propluvia(date, export = FALSE)
}
\arguments{
\item{date}{Date. Par défaut, la date du jour est utilisée. Si elle est renseignée, doit être au format ["\%Y-\%m-\%d"] (ex: 2022-07-23).}

\item{export}{Valeur logique (par défaut `FALSE`). Si `TRUE`, créer une couche au format [`.gpkg`] dans le répertoire de travail.}
}
\value{
data.frame au format `sf`. Contient les geometry des zones d'alerte.
}
\description{
[get_propluvia()] permet de télécharger les données Propluvia pour la France à une date donnée (par défaut la date du jour).
Les données contiennent la géométrie des zones d'alertes.
L'argument `export` de la fonction permet de sauvegarder les données au format [`.gpkg`] pour utilisation SIG.
Les données mises en forme contiennent une colonne avec le code couleur des zones d'alerte.
}
\examples{
\dontrun{
data_jour <- get_propluvia(export = FALSE)
}
}
