% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_propluvia_dpt.R
\encoding{UTF-8}
\name{get_propluvia_dpt}
\alias{get_propluvia_dpt}
\title{Télécharge les données Propluvia pour un département et à date du jour (ou spécifiée)}
\usage{
get_propluvia_dpt(dpt, date, export = FALSE)
}
\arguments{
\item{dpt}{Numérique. Correspond au numéro de département, par ex. 14 pour le Calvados.}

\item{date}{Date. Par défaut, la date du jour est utilisée. Si elle est renseignée, doit être au format ["\%Y-\%m-\%d"] (ex: 2022-07-23).}

\item{export}{Valeur logique (par défaut `FALSE`). Si `TRUE`, créer une couche au format [`.gpkg`] dans le répertoire de travail.}
}
\value{
data.frame au format `sf`. Contient les informations relatives aux arrêtés en vigueur dans le département (numéro, date de signature, date de validité).
    Les niveaux de restriction avec leurs dénominations (Vigilance, Alerte, Alerte renforcée, Crise) et le code couleur associé.
    Les noms des zones d'alertes.
    Les types de zones : superficielles (`SUP`) et/ou souterraines (`SOU`).
    Les geometries des zones d'alerte (multipolygones au format WGS84, crs 4326).
}
\description{
[get_propluvia_dpt()] permet de télécharger les données Propluvia pour un département à une date donnée (par défaut la date du jour).
Les données contiennent la géométrie des zones d'alertes, et les infos sur les arrêtés en vigueur dans le département.
L'argument `export` de la fonction permet de sauvegarder les données au format [`.gpkg`] pour utilisation SIG.
Les données mises en forme contiennent également une colonne avec le code couleur des zones d'alerte.
}
\examples{
\dontrun{
data_jour_dpt14 <- get_propluvia_dpt(dpt = 27,export = FALSE)
}
}
